requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.3.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
multidict==6.1.0
numpy==1.24.4
openai==1.82.1
orjson==3.10.15
packaging==24.2
propcache==0.2.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==8.5.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
yarl==1.15.2
